Moved sample filter object to src dir.
